feat: implement high-fidelity After Effects rendering parity (Tier 1)
- Modularized blending system with 38 AE-compatible modes
- Implemented AE-style keyframe easing with speed/influence handles
- Optimized Track Mattes using Rec.709 Luma math and raw pixel access
- Integrated temporal sub-sampling for Motion Blur with per-layer toggle
- Normalized color pipeline to 0..1 to harden effects (Glow, Blur, Shadow)

+++ src/media/video_decoder.cpp
@@ -58,10 +58,10 @@ renderer2d::SurfaceRGBA surface_from_rgba_buffer(const std::vector<std::uint8_t>
                 static_cast<std::uint32_t>(x),
                 static_cast<std::uint32_t>(y),
                 renderer2d::Color{
-                    buffer[index + 0U],
-                    buffer[index + 1U],
-                    buffer[index + 2U],
-                    buffer[index + 3U]
+                    static_cast<float>(buffer[index + 0U]),
+                    static_cast<float>(buffer[index + 1U]),
+                    static_cast<float>(buffer[index + 2U]),
+                    static_cast<float>(buffer[index + 3U])
                 });
         }
     }
